Assign platform=canvas to canvas resources
